+++ Oticon Genie 2 2025 Checker.py
@@ -111,7 +111,7 @@ while updaterRetries > 0:
         # This is required to get update for newer (2025+) versions.
         # Software names set by b:Name
         # Version numbers set by b:Major . b:Minor . b:Build . b:Revision
-        osVer = config.get('Oticon', 'OS', fallback='Microsoft Windows NT 10.0.22621.0')
+        osVer = config.get('General', 'OS', fallback='Microsoft Windows NT 10.0.22621.0')
         baseVer = config.get('Oticon', 'Version', fallback='20.22.95.0').split('.')
         updrVer = config.get('Oticon', 'Updater', fallback='27.2.19.0').split('.')
         baseVerMajor = baseVer[0]